Move WebSocketContext to it's own file
Keep it as child class (partial connection class) so that it can invoke events and private methods on the connection.
